Abstract getReactInstance to utils
