+++ packages/codelift/utils/getReactInstance.ts
@@ -1,0 +1,13 @@
+export const getReactInstance = (element: HTMLElement) => {
+  if ("_reactRootContainer" in element) {
+    // @ts-ignore Property '_reactRootContainer' does not exist on type 'never'.ts(2339)
+    return element._reactRootContainer._internalRoot.current.child;
+  }
+
+  for (const key in element) {
+    if (key.startsWith("__reactInternalInstance$")) {
+      // @ts-ignore No index signature with a parameter of type 'string' was found on type 'HTMLElement'.ts(7053)
+      return element[key];
+    }
+  }
+};
